chore : api folder 용도에 맞게 구분
